+++ package.json
@@ -12,7 +12,9 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "sass": "^1.77.4"
+    "sass": "^1.77.4",
+    "socket.io-client": "^4.7.5",
+    "websocket": "^1.0.35"
   },
   "devDependencies": {
     "@types/react": "^18.2.66",
