+++ src/components/Window/Window.jsx
@@ -1,21 +1,13 @@
-import { useEffect, useState, useRef } from 'react';
+import { useState } from 'react';
+
 import Calculator from '../Calculator/Calculator';
+
 import './Window.scss';
 
-const DraggableWindow = () => {
+const DraggableWindow = ({ maxX, maxY }) => {
   const [isDragging, setIsDragging] = useState(false);
   const [position, setPosition] = useState({ x: 0, y: 0 });
   const [initialPosition, setInitialPosition] = useState({ x: 0, y: 0 });
-  const [maxXY, setMaxXY] = useState({ maxX: 0, maxY: 0 });
-
-  const screenRef = useRef(null);
-
-  useEffect(() => {
-    if (screenRef.current) {
-      const screenBounds = screenRef.current.getBoundingClientRect();
-      setMaxXY({ maxX: screenBounds.width, maxY: screenBounds.height });
-    }
-  }, [screenRef]);
 
   const handleMouseDown = (e) => {
     setIsDragging(true);
@@ -35,16 +27,17 @@ const DraggableWindow = () => {
     let newX = event.clientX - initialPosition.x;
     let newY = event.clientY - initialPosition.y;
 
+    // Constrain within bounds
     if (newX < 0) {
       newX = 0;
-    } else if (newX > maxXY.maxX) {
-      newX = maxXY.maxX;
+    } else if (newX > maxX) {
+      newX = maxX;
     }
 
     if (newY < 0) {
       newY = 0;
-    } else if (newY > maxXY.maxY) {
-      newY = maxXY.maxY;
+    } else if (newY > maxY) {
+      newY = maxY;
     }
 
     setPosition({
@@ -53,14 +46,6 @@ const DraggableWindow = () => {
     });
   };
 
-  // const handleWindowResize = () => {
-  //   if (screenRef.current) {
-  //     const screenBounds = screenRef.current.getBoundingClientRect();
-  //     setMaxXY({ maxX: screenBounds.width, maxY: screenBounds.height });
-  //   }
-  // };
-
-  // window.addEventListener('resize', handleWindowResize);
   return (
     <div
       className="draggable-window"
@@ -72,7 +57,6 @@ const DraggableWindow = () => {
       onMouseDown={handleMouseDown}
       onMouseUp={handleMouseUp}
       onMouseMove={handleMouseMove}
-      ref={screenRef} // Assign the ref here
     >
       <div className="draggable-window_title-bar">
         <button>x</button>
